Moves creation of jobs and folders into its own class JobsBuilder.
That way it is kept track automatically of which folder the parent
folder of a job is.
Users do not have to specify the parent-hierarchy over over again.
Furthermore this enables to share the outer scope later on, without
affecting existing jobs.

This is a prerequisite of JENKINS-177, since there the JENKINS_HOME
variable needs to be accessed.
Yet JENKINS_HOME is only available in the outer scope.