JENKINS-171 Uses reference repositories for Catroid and Paintroid.
git allows to specify local reference repositories for its clone command.
This can speed-up cloning drastically, since this local cache reduces
the amount of data that has to be fetched from the web.

To make this as easy to maintain as possible the behavior is transparent.
That means there is a cache for each product on each node accessible via
the same relative path.

Furthermore there are jobs that keep the cache up to date on all the nodes.
When a new node is added it has to be specified in the git_cache.groovy
file to also have a cache ready.