Removes feature to exclude tests on Jenkins.
This feature was supposed to work around the issue that most tests do not work.
Yet as it turned out later the regular target for tests should not be used ...
well because it includes many outdated tests. ;)
With new targets in place the issue of having hundreds of failing tests
is not present anymore.