JENKINS-183 Cleanup workspace instead of deletion
Previously right before every build the workspace
got deleted.

To avoid loading the repository on every commit
from upstream, a git-cache was introduced to
improve cloning performance (JENKINS-171).

Now this change improves the situation again.
Due to the fact that the workspace is also the
root of the git-repo we can simply use the
cleanBeforeCheckout-option of git-scm to reset
the workspace without remove and cloning.

This also means, that the git-cache-jobs and the
referencing to them can be completely removed.