JENKINS-222: Pipeline does not support concurrentBuild
Remove call from generic JobBuilder, where it was used
inside the nightly block.
First there is no need to imply concurrentBuild(false) if
a jobs shall also run on a daily schedule.
Second by default a job is not marked as concurrent if no
call to concurrentBuild is done.
Still explicitly state the line in the SeedJob, as it was
the only freestyle-job using nightly and the job which
definitely will cause troubles if called concurrently,
should the default behaviour change.