JENKINS-261 Only scan the repository once per day.
First the webhooks work again.
Second rescanning leads to rebuilding every PR in case the target branch,
e.g. develop, changed.
This leads to a potential effort of O(n^2).
Doing this at most once per day ensures that the build results for the PRs
are still up-to-date but the incurred effort is not too high.