JENKINS-286 Build the seed job every Sunday before the nightlies.
When the seed job is build the multibranch jobs start scanning the branches
for changes, potentially leading to rebuilds.

When this scanning is performed should depend on the settings of the multibranch
job and not the seed job.
Running the seed job once per week strikes a trade-off between clean job
configurations and interaction with the multibranch jobs.